Refactor ExpectimaxPlayer to use uint64_t for scoring and probability calculations
- Change scoring methods from double to uint64_t for performance
- Modify probability weighting from floating-point to integer-based scaling
- Update method signatures in both implementation and header files
- Simplify score calculation and probability representation

+++ cpp/src/expectimax_player.hpp
@@ -35,9 +35,9 @@ private:
     EvaluationFunction evalFn;
     std::chrono::steady_clock::time_point startTime;
 
-    double expectimax(uint64_t state, int depth, bool isMax);
-    double chanceNode(uint64_t state, int depth);
-    double maxNode(uint64_t state, int depth);
+    uint64_t expectimax(uint64_t state, int depth, bool isMax);
+    uint64_t chanceNode(uint64_t state, int depth);
+    uint64_t maxNode(uint64_t state, int depth);
     bool shouldTimeOut() const;
     int getAdaptiveDepth(uint64_t state) const;
 };
